refactor(chat-agent): co-locate tests with source files
